requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
faiss-cpu
python-dotenv
tiktoken
supabase